+++ golang/lib/bulk_command_execution/v0_bulk_command_api/v0_bulk_commands.go
@@ -119,10 +119,19 @@ func (visitor *cmdArgDeserializingVisitor) VisitExecCommand() error {
 	return nil
 }
 
-func (visitor *cmdArgDeserializingVisitor) VisitWaitForEndpointAvailability() error {
-	args := &kurtosis_core_rpc_api_bindings.WaitForEndpointAvailabilityArgs{}
+func (visitor *cmdArgDeserializingVisitor) VisitWaitForHttpGetEndpointAvailability() error {
+	args := &kurtosis_core_rpc_api_bindings.WaitForHttpGetEndpointAvailabilityArgs{}
 	if err := json.Unmarshal(visitor.bytesToDeserialize, args); err != nil {
-		return stacktrace.Propagate(err, "An error occurred deserializing the endpoint availability-waiting args")
+		return stacktrace.Propagate(err, "An error occurred deserializing the endpoint availability-waiting-http-get args")
+	}
+	visitor.deserializedCommandArgsPtr = args
+	return nil
+}
+
+func (visitor *cmdArgDeserializingVisitor) VisitWaitForHttpPostEndpointAvailability() error {
+	args := &kurtosis_core_rpc_api_bindings.WaitForHttpPostEndpointAvailabilityArgs{}
+	if err := json.Unmarshal(visitor.bytesToDeserialize, args); err != nil {
+		return stacktrace.Propagate(err, "An error occurred deserializing the endpoint availability-waiting-http-post args")
 	}
 	visitor.deserializedCommandArgsPtr = args
 	return nil